fixing build and tests after projection changes, refs #21, #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@8107 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/geom/GeoConvHelper.cpp
@@ -28,6 +28,7 @@
 #endif
 
 #include <map>
+#include <cmath>
 #include "GeoConvHelper.h"
 #include <utils/common/MsgHandler.h>
 #include <utils/common/ToString.h>
@@ -131,7 +132,7 @@ GeoConvHelper::init(const std::string &proj,
                     const int shift,
                     bool inverse) {
     myProjString = proj;
-    myGeoScale = pow(10, -shift);
+    myGeoScale = pow(10, (double)-shift);
     myUseInverseProjection = inverse;
     close();
     myBaseFound = false;
